Fix _queue_increase_capacity to copy correct slice
This function was previously copying a large, overlapping
range of data when the queue resized. The function caused
program to fail on out of bounds slices if the queue was
pop_front-ed from (offset > 0) and resized on full.

New code copies the range [0:offset] to [end:end+offset] where
end is the previous size of the queue buffer