Set .Reuse_Address on listening TCP sockets
On Windows, you can set Exclusive_Addr_Use, and it will prevent
hijacking of the socket address, IIUC.
It also appears to mean that you can bypass the cooldown period of the
socket if the program quits without closing the socket.

However, on Linux this option doesn't exist.
It seems that Reuse_Address on the listen socket allows you to bypass
the cooldown period, as per - but we should obviously verify
that the hijacking problem--is in fact--not a problem here.