Fix UDP error handling of payload truncation in recv_udp()
On Linux, when the receive buffer is not large enough to hold the entire
UDP payload, the remainder of the message is silently truncated.

But, if you pass the MSG_TRUNC flag, the syscall returns the total size
of the payload, rather than the number of bytes it wrote into your
buffer.

We take advantage of this here to detect this situation, and return the
.Buffer_Too_Small error.