A better user formatter interface, that allows many users/libraries to interface to it.
I have made it so that there is no need for the user to make a map, and that many different parts of the code can contribute to the user map without fighting each other about who gets to call set_user_formatters. Also it will no longer return an error, instead the user can check if a formatter is already set.