[thread] Fix temp allocator not being thread safe when creating new threads
The default temp allocator is thread safe, but it achieves this via its global state being @(thread_local).
This means that &this_global_state must be taken __inside the thread__ that is going to be using it.

Previously, on Windows it was being set to the caller's context, which includes the temp allocator pointer.
It was always setting the init_context, which is only intended for when the user specifically wants to use a custom context, and not the default
one.
This meant that, __by default__, you'd get a temp allocator that WASN'T thread-safe!

It also wasn't entirely clear how the user could actually set the initial context with `thread.create()`, or how it was meant
to be used. This now has `thread.create()` take it as a default parameter as well, to match all the other thread-start procs.