[core:thread] Fix `run` procedure
This commit fixes the behavior of `run`. Previously it would do `destroy(t)`, which would attempt to join the thread unto itself, which made no sense, thus it just deadlocked, and `free(t)` was never called.

This change adds `Auto_Cleanup` flag that is set when `run` procedure is called. Thread frees its memory at the end of its life if it finds that flag.