add frame pointer attribute to procs in debug mode
The frame pointer attribute is an important attribute when debugging and
traversing the call stack of MacOS programs. Without the frame pointer
programs are unable to traverse the stack further than the most recent
stack frame.

My main motivation for this change is adding MacOS support to [my
backtrace package](https://github.com/laytan/obacktracing).

I am not sure of the implications of the attribute on other platforms.