Corrects ColorToInt signature to return int instead of uint
The current C definition of ColorToInt in the raylib.h header itself
is: `RLAPI int ColorToInt(Color color);`

Though it feels improper (and perhaps relies on undefined behavior?),
the raylib implementation does indeed cast an unsigned integer to
a signed one:

https://github.com/raysan5/raylib/blob/7fab03c0b4f079a6450f03d82ca17f2e18be3477/src/rtextures.c#L4893
```
int ColorToInt(Color color)
{
    int result = 0;

    result = (int)(((unsigned int)color.r << 24) |
                   ((unsigned int)color.g << 16) |
                   ((unsigned int)color.b << 8) |
                    (unsigned int)color.a);

    return result;
}
```

This is unfortunately a breaking change for users of this package. I'm
unsure of whether a different approach would be better. This incorrect
signature made it impossible, from what I could tell, to be able
to set default control colors in raygui.

```
    rl.GuiSetStyle(.DEFAULT, i32(rl.GuiDefaultProperty.LINE_COLOR), rl.ColorToInt(rl.WHITE))
```

With the previously incorrect signature, I couldn't find a way to cast a
raylib color to i32 in a forced way. Perhaps my own Odin knowledge shortcoming.